move instructions into its own component
