+++ src/frontend/src/common/components/library/data_subview/components/EditSamplesConfirmationModal/style.ts
@@ -1,13 +1,6 @@
 import styled from "@emotion/styled";
 import IconButton from "@material-ui/core/IconButton";
-import {
-  Button,
-  fontBodyS,
-  fontBodyXs,
-  fontBodyXxs,
-  getColors,
-  getSpaces,
-} from "czifui";
+import { fontBodyS, fontBodyXs, getColors, getSpaces } from "czifui";
 
 export const StyledIconButton = styled(IconButton)`
   display: flex;
@@ -46,21 +39,6 @@ export const StyledSubTitle = styled.span`
     return `
       color: ${colors?.gray[600]};
       margin-bottom: ${spaces?.xl}px;
-    `;
-  }}
-`;
-
-export const StyledButton = styled(Button)`
-  ${fontBodyXxs}
-  ${(props) => {
-    const colors = getColors(props);
-    const spaces = getSpaces(props);
-    return `
-      padding-bottom: ${spaces?.s}px;
-      color: ${colors?.primary[500]};
-      &:hover {
-        background-color: transparent;
-      }
     `;
   }}
 `;
